+++ src/Handler/ErrorLogHandlerFactory.php
@@ -11,10 +11,10 @@ class ErrorLogHandlerFactory implements FactoryInterface
 {
     public function __invoke(array $options)
     {
-        $messageType = (int) $options['messageType'] ?? ErrorLogHandler::OPERATING_SYSTEM;
-        $level = $options['level'] ?? Logger::DEBUG;
-        $bubble = (boolean) $options['bubble'] ?? true;
-        $expandNewlines = (boolean) $options['expandNewlines'] ?? false;
+        $messageType    = (int)     ($options['messageType']    ?? ErrorLogHandler::OPERATING_SYSTEM);
+        $level          = (int)     ($options['level']          ?? Logger::DEBUG);
+        $bubble         = (boolean) ($options['bubble']         ?? true);
+        $expandNewlines = (boolean) ($options['expandNewlines'] ?? false);
 
         return new ErrorLogHandler($messageType, $level, $bubble, $expandNewlines);
     }
