+++ src/Handler/NativeMailerHandlerFactory.php
@@ -11,12 +11,12 @@ class NativeMailerHandlerFactory implements FactoryInterface
 {
     public function __invoke(array $options)
     {
-        $toEmail = (array) $options['to'] ?? [];
-        $subject = (string) $options['subject'] ?? true;
-        $fromEmail = (string) $options['from'] ?? '';
-        $level = $options['level'] ?? Logger::DEBUG;
-        $bubble = (boolean) $options['bubble'] ?? true;
-        $maxColumnWidth = (int) $options['maxColumnWidth'] ?? 70;
+        $toEmail        = (array)   ($options['to']             ?? []);
+        $subject        = (string)  ($options['subject']        ?? true);
+        $fromEmail      = (string)  ($options['from']           ?? '');
+        $level          = (int)     ($options['level']          ?? Logger::DEBUG);
+        $bubble         = (boolean) ($options['bubble']         ?? true);
+        $maxColumnWidth = (int)     ($options['maxColumnWidth'] ?? 70);
 
         return new NativeMailerHandler($toEmail, $subject, $fromEmail, $level, $bubble, $maxColumnWidth);
     }
